requirements.txt
PySide6
numpy
PyOpenGL
PyOpenGL_accelerate
